block users from disabling the please-reboot notification
Statically grant Updater permission for "fixed" notifications;
set most notifications to be blockable by users, but set
notifications on the REBOOT channel to be non-blockable.

Some users have gotten "stuck" by disabling the notification
to reboot the system to install a new release.  Updater can't
download and prepare a later release until the previously-prepared
release has been installed, so it is important to avoid losing
the please-reboot notification.